+++ packages/thread/src/export.ts
@@ -1,1 +1,2 @@
 export * from './worker.js';
+export * from './workerTypes.js';