+++ src/types/export.ts
@@ -1,3 +1,0 @@
-export * from '../frontendTypes.js';
-export * from '../matterbridgeTypes.js';
-export * from '../matterbridgePlatformTypes.js';